Move content to navbar module
